+++ HISTORY.md
@@ -5,6 +5,7 @@
  * Fixed; Replaced `helpers.jsonRequest()` with SuperAgent due to a bug in `JQuery $.ajax()`.
  * Added; Sage and bump functions for posts.
  * Added; Delete function for threads and posts.
+ * Changed; Client-side router to use hashbang urls (`#!`).
 
 0.0.2 / 2012-09-10
 ------------------